requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-sslcommerz==1.0.0a0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
pillow==10.4.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pyotp==2.9.0
pyparsing==3.1.2
python-decouple==3.8
pytz==2024.1
requests==2.31.0
rsa==4.9
sqlparse==0.5.0
sslcommerz-lib==1.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
